rename GKOBase to GKOlduBase
