+++ GKOlduBase/GKOlduBase.C
@@ -26,7 +26,7 @@ License
 #include <ginkgo/ginkgo.hpp>
 #include <map>
 #include <type_traits>
-#include "GKOBase.H"
+#include "GKOlduBase.H"
